hidpp10: read_memory should take offset in bytes
The current API takes offset in 2-byte values. Taking a byte value
is more natural. We need to upgrade the type of offset to uint16_t
to handle values > 256.

+++ src/hidpp10.h
@@ -462,7 +462,7 @@ hidpp10_set_usb_refresh_rate(struct hidpp10_device *dev,
 int
 hidpp10_read_memory(struct hidpp10_device *dev,
 		    uint8_t page,
-		    uint8_t offset,
+		    uint16_t offset,
 		    uint8_t bytes[16]);
 
 /* -------------------------------------------------------------------------- */
